requirements.txt
mapproxy
PyYAML
